Consolidate CosmosRelay implementations into one place
