--- crates/cosmos/cosmos-relayer/src/impls/relay/component.rs
@@ -1,58 +1,0 @@
-use cgp_core::prelude::*;
-use cgp_core::{delegate_all, ErrorRaiserComponent, ErrorTypeComponent};
-use hermes_logging_components::traits::has_logger::{
-    GlobalLoggerGetterComponent, LoggerGetterComponent, LoggerTypeComponent,
-};
-use hermes_relayer_components::error::impls::retry::ReturnMaxRetry;
-use hermes_relayer_components::error::traits::retry::{
-    MaxErrorRetryGetterComponent, RetryableErrorComponent,
-};
-use hermes_relayer_components::relay::impls::packet_lock::ProvidePacketLockWithMutex;
-use hermes_relayer_components::relay::traits::packet_lock::PacketLockComponent;
-use hermes_relayer_components_extra::components::extra::closures::relay::auto_relayer::CanUseExtraAutoRelayer;
-use hermes_relayer_components_extra::components::extra::relay::{
-    ExtraRelayComponents, IsExtraRelayComponent,
-};
-use hermes_runtime::impls::types::runtime::ProvideHermesRuntime;
-use hermes_runtime_components::traits::runtime::RuntimeTypeComponent;
-
-use crate::contexts::logger::ProvideCosmosLogger;
-use crate::contexts::relay::CosmosRelay;
-use crate::impls::error::HandleCosmosError;
-
-pub struct CosmosRelayComponents;
-
-delegate_components! {
-    CosmosRelayComponents {
-        [
-            ErrorTypeComponent,
-            ErrorRaiserComponent,
-            RetryableErrorComponent,
-        ]:
-            HandleCosmosError,
-        RuntimeTypeComponent:
-            ProvideHermesRuntime,
-        [
-            LoggerTypeComponent,
-            LoggerGetterComponent,
-            GlobalLoggerGetterComponent,
-        ]:
-            ProvideCosmosLogger,
-        MaxErrorRetryGetterComponent:
-            ReturnMaxRetry<3>,
-        PacketLockComponent:
-            ProvidePacketLockWithMutex,
-    }
-}
-
-delegate_all!(
-    IsExtraRelayComponent,
-    ExtraRelayComponents,
-    CosmosRelayComponents,
-);
-
-impl HasComponents for CosmosRelay {
-    type Components = CosmosRelayComponents;
-}
-
-impl CanUseExtraAutoRelayer for CosmosRelay {}
