--- crates/cosmos/cosmos-relayer/src/impls/relay/mod.rs
@@ -1,5 +1,0 @@
-pub mod batch;
-pub mod component;
-pub mod packet_filter;
-pub mod packet_lock;
-pub mod types;
